// src/service-api/generated/models/reportsActivityGranularityDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type ReportsActivityGranularityDto = typeof ReportsActivityGranularityDto[keyof typeof ReportsActivityGranularityDto];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportsActivityGranularityDto = {
  day: 'day',
  week: 'week',
} as const;